test(daemon): split the two replay test files over the size tripwire
No daemon test file should sit over the 1,000-line tripwire, so the size
ratchet can fail on new size, not only on growth. Split the two files PR
#2417 left over into sibling `<base>-<concern>.test.ts` files:

- session-replay-divergence.test.ts -> suggestions, chrome-filter, overlay,
  capture-policy, observation (13 cases).
- session-replay-runtime-maestro.test.ts -> dispatch, run-script, targeting,
  swipe, keyboard, run-flow, shell-artifacts (43 cases).

Shared setup moves to a dedicated fixtures module per family:
- session-replay-divergence.fixtures.ts holds the device-resolution and
  snapshot-interactor mocks and the shared reset, and re-exports the SUTs so
  siblings load them under the mocks. The freshness-retry `sleep` stub lives in
  a separate retry fixtures module the two retry siblings import first, so it
  cannot no-op the retry delay in siblings that never retry.
- session-replay-runtime-maestro.fixtures.ts declares the Maestro
  device-resolution vi.mock once, so the split does not copy it per file.

The diff is moves, import paths and fixture extraction. The case count under
src/daemon/replay/internal/__tests__ is unchanged (13 + 43 = 56); no test is
deleted or skipped, and assertions are identical.

Closes #2446
