+++ src/daemon/replay/internal/__tests__/session-replay-runtime-maestro-shell-artifacts.test.ts
@@ -1,0 +1,112 @@
+import { maestroReplayFixture } from './session-replay-runtime-maestro.fixtures.ts';
+import assert from 'node:assert/strict';
+import fs from 'node:fs';
+import path from 'node:path';
+import { test } from 'vitest';
+import { mkdtempForTestSync } from '../../../../__tests__/test-utils/tmp-dir.ts';
+import { SessionStore } from '../../../session-store.ts';
+import { replayScriptSourceBundleFor } from '../../../../__tests__/test-utils/replay-script-source.ts';
+
+const { runReplayFixture, runReplayForTest } = maestroReplayFixture;
+
+test('runReplayCommand reads shell env from request (client-collected), not daemon process.env', async () => {
+  // Ensure the daemon's own process.env does NOT contain AD_VAR_APP.
+  assert.equal(process.env.AD_VAR_APP, undefined);
+  const { response, calls } = await runReplayFixture({
+    label: 'shell',
+    script: 'context platform=android\nopen ${APP}\n',
+    // Client-collected shell env; still uses the raw AD_VAR_* prefix.
+    flags: { replayShellEnv: { AD_VAR_APP: 'client-shell-app' } },
+  });
+  assert.equal(response.ok, true);
+  assert.deepEqual(calls[0]?.positionals, ['client-shell-app']);
+});
+
+test('runReplayCommand falls back to process.env when request omits replayShellEnv', async () => {
+  const previous = process.env.AD_VAR_APP;
+  process.env.AD_VAR_APP = 'daemon-env-app';
+  try {
+    const { response, calls } = await runReplayFixture({
+      label: 'shell-fallback',
+      script: 'context platform=android\nopen ${APP}\n',
+    });
+    assert.equal(response.ok, true);
+    assert.deepEqual(calls[0]?.positionals, ['daemon-env-app']);
+  } finally {
+    if (previous === undefined) delete process.env.AD_VAR_APP;
+    else process.env.AD_VAR_APP = previous;
+  }
+});
+
+test('runReplayCommand writes per-action timing events to active trace', async () => {
+  const root = mkdtempForTestSync('agent-device-replay-trace-');
+  const scriptPath = path.join(root, 'flow.ad');
+  const tracePath = path.join(root, 'trace.ndjson');
+  fs.writeFileSync(scriptPath, 'context platform=ios\nclick id="submit"\nwait "Done" 5000\n');
+  fs.writeFileSync(tracePath, '');
+
+  const sessionStore = new SessionStore(path.join(root, 'state'));
+  sessionStore.set('s', {
+    name: 's',
+    device: {
+      platform: 'apple',
+      id: 'sim-1',
+      name: 'iPhone',
+      kind: 'simulator',
+      booted: true,
+    },
+    createdAt: Date.now(),
+    trace: { outPath: tracePath, startedAt: Date.now() },
+    actions: [],
+  });
+
+  const response = await runReplayForTest({
+    req: {
+      token: 't',
+      session: 's',
+      command: 'replay',
+      positionals: [scriptPath],
+      flags: { replayScriptSource: replayScriptSourceBundleFor(scriptPath) },
+      meta: { cwd: root },
+    },
+    sessionName: 's',
+    logPath: path.join(root, 'log'),
+    sessionStore,
+    invoke: async (req) => ({
+      ok: true,
+      data:
+        req.command === 'click'
+          ? { timing: { totalDurationMs: 12, internal: { ignored: true } } }
+          : {},
+    }),
+  });
+
+  assert.equal(response.ok, true);
+  const events = fs
+    .readFileSync(tracePath, 'utf8')
+    .trim()
+    .split('\n')
+    .map((line) => JSON.parse(line) as Record<string, unknown>);
+  assert.deepEqual(
+    events.map((event) => [event.type, event.step, event.command]),
+    [
+      ['replay_action_start', 1, 'click'],
+      ['replay_action_stop', 1, 'click'],
+      ['replay_action_start', 2, 'wait'],
+      ['replay_action_stop', 2, 'wait'],
+    ],
+  );
+  assert.equal(typeof events[1]?.durationMs, 'number');
+  assert.deepEqual(events[1]?.resultTiming, { totalDurationMs: 12 });
+});
+
+test('AD_ARTIFACTS resolves to per-attempt dir when artifactsDir flag is set by the test runner', async () => {
+  const attemptDir = '/tmp/agent-device-replay-artifacts-stub/run-x/flow/attempt-1';
+  const { response, calls } = await runReplayFixture({
+    label: 'artifacts',
+    script: 'context platform=android\nscreenshot "${AD_ARTIFACTS}/shot.png"\n',
+    flags: { artifactsDir: attemptDir },
+  });
+  assert.equal(response.ok, true);
+  assert.deepEqual(calls[0]?.positionals, [`${attemptDir}/shot.png`]);
+});
